+++ src/query/v2/multiframe.hpp
@@ -30,7 +30,7 @@ class MultiFrame {
   friend class ValidFramesReader;
   friend class InvalidFramesPopulator;
 
-  MultiFrame(int64_t size_of_frame, size_t number_of_frames, utils::MemoryResource *execution_memory);
+  MultiFrame(size_t size_of_frame, size_t number_of_frames, utils::MemoryResource *execution_memory);
   ~MultiFrame() = default;
 
   MultiFrame(const MultiFrame &other);
